Restructe back to one file and new division of tests
