+++ axlearn/common/utils_neuron.py
@@ -1,35 +1,46 @@
-# Copyright © 2024 Apple Inc.
-#
-# Some of the code in this file is adapted from:
-#
-# tensorflow/lingvo:
+
 # Copyright 2018 The TensorFlow Authors. All Rights Reserved.
 # Licensed under the Apache License, Version 2.0 (the "License").
 #
 # google/praxis:
 # Copyright 2022 The Pax Authors.
 # Licensed under the Apache License, Version 2.0 (the "License").
-"""Unit Test for mixture_of_experts.py"""
+"""Utils for tests for mixture_of_experts.py"""
 from functools import partial
 from itertools import product
 
 import jax
 import jax.numpy as jnp
-from absl.testing import absltest, parameterized
 
 from axlearn.common.mixture_of_experts import (
     Top2Gating,
-    TopKGating,
     TransformerFeedForwardMoE,
     TopKGatingGather,
 )
-from axlearn.common.module import functional as F
-from axlearn.common.test_utils import TestCase
 
 jax.config.update('jax_platform_name', 'cpu')
 
 MODULE_UNIT_TEST_ATOL=1e-6
 MODULE_UNIT_TEST_RTOL=1e-3
+
+# Grid space for testing
+# batchs =            [1, 4]
+# seqs =              [16, 128]
+# input_dims =        [64]
+# hidden_dims =       [128]
+# num_experts =       [2, 8]
+# num_groups =        [1, 4]
+# outer_batches =     [1, 2]
+# expert_capacities = [2, 1000]
+
+batchs =            [1]
+seqs =              [16]
+input_dims =        [64]
+hidden_dims =       [128]
+num_experts =       [2]
+num_groups =        [1]
+outer_batches =     [1]
+expert_capacities = [2]
 
 class ModuleConfig():
     def __init__(self, module = None, device = "cpu", mesh = (1,)):
@@ -60,7 +71,7 @@ class TestConfig():
 
         self.test_state = self.test_layer.initialize_parameters_recursively(prng_key=jax.random.PRNGKey(123))
         self.golden_state = self.test_state
-    
+
 def _topkgather_to_topk(output, expert_cap):
     tok_perm_idx, expert_index, exp_aff_mask = output.combine_tensor
 
@@ -162,57 +173,81 @@ class TestConfigBuilder:
             "train_capacity_factor": self.params["train_capacity_factor"]
         }
     
-    def build_test_configs_moe(self):
-        return [
-            TestConfig(
-                setup=[
-                    self.build_moe_topkgather_setup(),
-                    self.build_moe_top2_setup()
-                ],
-                test=ModuleConfig(TransformerFeedForwardMoE, "cpu"),
-                golden=ModuleConfig(TransformerFeedForwardMoE, "cpu"),
-                inputs=dict(inputs=jax.random.uniform(
-                    jax.random.PRNGKey(1),
-                    shape=(self.params["batch_size"], self.params["seq_len"], self.params["input_dim"])
-                )),
-                loss_fn=lambda x: x.mean(),
-                prefix="_unit_test"
-            ),
-        ]
-
-    def build_test_configs_topk(self):
+    def build_test_configs_integ(self):
 
         seq_len = (self.params["batch_size"]*self.params["seq_len"])//(self.params["outer_batch"] * self.params["num_groups"])
 
         return [
             TestConfig(
                 setup=[
+                    self.build_moe_topkgather_setup(),
+                    self.build_moe_topkgather_setup()
+                ],
+                test=ModuleConfig(TransformerFeedForwardMoE, "neuron"),
+                golden=ModuleConfig(TransformerFeedForwardMoE, "cpu"),
+                inputs=dict(inputs=jax.random.uniform(
+                    jax.random.PRNGKey(1),
+                    shape=(self.params["batch_size"], self.params["seq_len"], self.params["input_dim"])
+                )),
+                loss_fn=lambda x: x.mean(),
+                prefix="_moe"
+            ),
+            TestConfig(
+                setup=[
                     self.build_gating_setup(),
                     self.build_gating_setup()
                 ],
-                test=ModuleConfig(TopKGatingGather, "cpu"),
-                golden=ModuleConfig(TopKGating, "cpu"),
+                test=ModuleConfig(TopKGatingGather, "neuron"),
+                golden=ModuleConfig(TopKGatingGather, "cpu"),
+                inputs=dict(logits=jax.random.uniform(
+                    jax.random.PRNGKey(1),
+                    shape=(self.params["outer_batch"], self.params["num_groups"],
+                           seq_len, self.params["num_experts"])
+                )),
+                loss_fn=lambda x: x.load_balance_loss,
+                prefix="_gating"
+            ),
+        ]
+
+    def build_test_configs_unit(self):
+
+        seq_len = (self.params["batch_size"]*self.params["seq_len"])//(self.params["outer_batch"] * self.params["num_groups"])
+
+        return [
+            TestConfig(
+                setup=[
+                    self.build_moe_topkgather_setup(),
+                    self.build_moe_top2_setup()
+                ],
+                test=ModuleConfig(TransformerFeedForwardMoE),
+                golden=ModuleConfig(TransformerFeedForwardMoE),
+                inputs=dict(inputs=jax.random.uniform(
+                    jax.random.PRNGKey(1),
+                    shape=(self.params["batch_size"], self.params["seq_len"], self.params["input_dim"])
+                )),
+                loss_fn=lambda x: x.mean(),
+                prefix="_moe"
+            ),
+            TestConfig(
+                setup=[
+                    self.build_gating_setup(),
+                    self.build_gating_setup()
+                ],
+                test=ModuleConfig(TopKGatingGather),
+                golden=ModuleConfig(Top2Gating),
                 inputs=dict(logits=jax.random.uniform(
                     jax.random.PRNGKey(1),
                     shape=(self.params["outer_batch"], self.params["num_groups"],
                            seq_len, self.params["num_experts"])
                 )),
                 conv_output=partial(_topkgather_to_topk, expert_cap=self.params["expert_capacity"]),
-                prefix="_unit_test"
+                loss_fn=lambda x: x.load_balance_loss,
+                prefix="_gating"
             ),
         ]
     
-def _get_training_configs():
+def get_training_configs(is_unit: bool = False):
     builder = TestConfigBuilder()
-
-    batchs =            [1, 4]
-    seqs =              [16, 128]
-    input_dims =        [64]
-    hidden_dims =       [128]
-    num_experts =       [2, 8]
-    num_groups =        [1, 4]
-    outer_batches =     [1, 2]
-    expert_capacities = [2, 1000]
 
     test_configs = []
 
@@ -222,122 +257,22 @@ def _get_training_configs():
         if batch % out_batch != 0:
             continue
         
-        config = builder.reset().with_dimensions(batch, seq, input_dim).with_expert_settings(
+        config = builder.reset()
+        config = config.with_dimensions(batch, seq, input_dim)
+        config = config.with_expert_settings(
             hidden_dim,
             out_batch,
             n_groups,
             n_experts,
             capacity,
             train_capacity_factor=None
-        ).build_test_configs_moe()
+        )
+        if is_unit:
+            config = config.build_test_configs_unit()
+        else:
+            config = config.build_test_configs_integ()
+
         name = f"MoE_b{batch}_s{seq}_i{input_dim}_h{hidden_dim}_e{n_experts}_g{n_groups}_ob{out_batch}_ec{capacity}"
         test_configs.extend([(name + cfg.prefix, cfg) for cfg in config])
-        
-        config = builder.build_test_configs_topk()
-        name = f"Gating_b{batch}_s{seq}_i{input_dim}_h{hidden_dim}_e{n_experts}_g{n_groups}_ob{out_batch}_ec{capacity}"
-        test_configs.extend([(name + cfg.prefix, cfg) for cfg in config])
 
     return test_configs
-
-def _get_training_configs_bwd():
-    builder = TestConfigBuilder()
-
-    batchs =            [1, 4]
-    seqs =              [16, 128]
-    input_dims =        [64]
-    hidden_dims =       [128]
-    num_experts =       [2, 8]
-    num_groups =        [1, 4]
-    outer_batches =     [1, 2]
-    expert_capacities = [2, 1000]
-
-    test_configs = []
-
-    for (batch, seq, input_dim,  hidden_dim, n_experts, n_groups, out_batch, capacity) in product(
-         batchs, seqs, input_dims, hidden_dims, num_experts, num_groups, outer_batches, expert_capacities):
-        
-        if batch % out_batch != 0:
-            continue
-
-        config = builder.reset().with_dimensions(batch, seq, input_dim).with_expert_settings(
-                    hidden_dim,
-                    out_batch,
-                    n_groups,
-                    n_experts,
-                    capacity,
-                    train_capacity_factor=None
-                ).build_test_configs_moe()
-        name = f"b{batch}_s{seq}_i{input_dim}_h{hidden_dim}_e{n_experts}_g{n_groups}_ob{out_batch}_ec{capacity}"
-        test_configs.extend([(name + cfg.prefix, cfg) for cfg in config])
-
-    return test_configs
-
-# pylint: disable=no-self-use,protected-access
-class TestImplCorrectness(TestCase):
-
-    def _fwd_call(self, layer, state, inputs):
-        return F(
-                layer,
-                is_training=True,
-                prng_key=jax.random.PRNGKey(123),
-                state=state,
-                inputs=inputs,
-        )
-
-    @parameterized.named_parameters(_get_training_configs())
-    def test_fwd_correctness(self, cfg: TestConfig):
-
-        @partial(jax.jit, backend=cfg.test.device)
-        def test_fwd_call():
-            test_output, _ = self._fwd_call(cfg.test_layer, cfg.test_state, cfg.inputs)
-            return test_output
-
-        @partial(jax.jit, backend=cfg.golden.device)
-        def golden_fwd_call():
-            golden_output, _ =  self._fwd_call(cfg.golden_layer, cfg.golden_state, cfg.inputs)
-            return golden_output
-        
-        test_output = test_fwd_call()
-        golden_output = golden_fwd_call()
-
-        if cfg.conv_output != None:
-            test_output = cfg.conv_output(test_output)
-        
-        # Transfer results to CPU before comparison
-        self.assertNestedAllClose(jax.device_get(test_output), jax.device_get(golden_output))
-
-    @parameterized.named_parameters(_get_training_configs_bwd())
-    def test_bwd_correctness(self, cfg: TestConfig):
-
-        @partial(jax.jit, backend=cfg.test.device)
-        def test_bwd_call(state):
-            def loss_fn(state):
-                test_output, _ = self._fwd_call(cfg.test_layer, state, cfg.inputs)
-                return cfg.loss_fn(test_output)
-            
-            loss, grads = jax.value_and_grad(loss_fn, has_aux=False)(state)
-            return loss, grads
-
-        @partial(jax.jit, backend=cfg.golden.device)
-        def golden_bwd_call(state):
-            def loss_fn(state):
-                golden_output, _ = self._fwd_call(cfg.golden_layer, state, cfg.inputs)
-                return cfg.loss_fn(golden_output)
-            
-            loss, grads = jax.value_and_grad(loss_fn, has_aux=False)(state)
-            return loss, grads
-
-        test_loss, test_grads = test_bwd_call(cfg.test_state)
-        golden_loss, golden_grads = golden_bwd_call(cfg.golden_state)
-
-        # Transfer results to CPU before comparison
-        test_loss = jax.tree_map(jax.device_get, test_loss)
-        golden_loss = jax.tree_map(jax.device_get, golden_loss)
-        test_grads = jax.tree_map(jax.device_get, test_grads)
-        golden_grads = jax.tree_map(jax.device_get, golden_grads)
-        
-        self.assertNestedAllClose(test_loss, golden_loss)
-        self.assertNestedAllClose(test_grads, golden_grads)
-
-if __name__ == "__main__":
-    absltest.main()
